// frontend/src/plugins/pluginApi.generated.ts
export type PluginDataVariables = {
  storeId: string;
  pluginCode: string;
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    storeId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    relatedRecordId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    dataIdentifier?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    datetime?: {
    equalTo?: string | null;
    beforeOrEqualTo?: string | null;
    afterOrEqualTo?: string | null;
  } | null;
  } | null;
};

export type PluginDataResult = {
  pluginData: ({
  totalCount: number;
  nodes: Array<{
  id: string;
  pluginCode: string;
  relatedRecordId: string | null;
  dataIdentifier: string;
  storeId: string | null;
  data: string;
  datetime: string | null;
}>;
});
};

export const PluginData = {
  query: "query pluginData($storeId: String!, $pluginCode: String!, $filter: PluginDataFilterInput) {\n  pluginData(storeId: $storeId, pluginCode: $pluginCode, filter: $filter) {\n    ... on PluginDataConnector {\n      totalCount\n      nodes {\n        id\n        pluginCode\n        relatedRecordId\n        dataIdentifier\n        storeId\n        data\n        datetime\n      }\n    }\n  }\n}",
} as TypedDocument<PluginDataResult, PluginDataVariables>;

export type InsertPluginDataVariables = {
  storeId: string;
  input: {
    id: string;
    storeId?: string | null;
    pluginCode: string;
    relatedRecordId?: string | null;
    dataIdentifier: string;
    data: string;
    datetime?: string | null;
  };
};

export type InsertPluginDataResult = {
  insertPluginData: ({
  id: string;
  data: string;
  dataIdentifier: string;
  relatedRecordId: string | null;
  datetime: string | null;
});
};

export const InsertPluginData = {
  query: "mutation insertPluginData($storeId: String!, $input: InsertPluginDataInput!) {\n  insertPluginData(storeId: $storeId, input: $input) {\n    ... on PluginDataNode {\n      id\n      data\n      dataIdentifier\n      relatedRecordId\n      datetime\n    }\n  }\n}",
} as TypedDocument<InsertPluginDataResult, InsertPluginDataVariables>;

export type UpdatePluginDataVariables = {
  storeId: string;
  input: {
    id: string;
    storeId?: string | null;
    pluginCode: string;
    relatedRecordId?: string | null;
    dataIdentifier: string;
    data: string;
    datetime?: string | null;
  };
};

export type UpdatePluginDataResult = {
  updatePluginData: ({
  id: string;
  data: string;
  dataIdentifier: string;
  relatedRecordId: string | null;
  datetime: string | null;
});
};

export const UpdatePluginData = {
  query: "mutation updatePluginData($storeId: String!, $input: UpdatePluginDataInput!) {\n  updatePluginData(storeId: $storeId, input: $input) {\n    ... on PluginDataNode {\n      id\n      data\n      dataIdentifier\n      relatedRecordId\n      datetime\n    }\n  }\n}",
} as TypedDocument<UpdatePluginDataResult, UpdatePluginDataVariables>;